requirements.txt
ultralytics
opencv-python
azure-cognitiveservices-speech
requests
